// src/lib/api/generated/models/getProduceListHasDelivery.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export type GetProduceListHasDelivery = typeof GetProduceListHasDelivery[keyof typeof GetProduceListHasDelivery];


export const GetProduceListHasDelivery = {
  true: 'true',
  false: 'false',
} as const;