// src/lib/api/generated/models/getProduceListIsSubscribable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export type GetProduceListIsSubscribable = typeof GetProduceListIsSubscribable[keyof typeof GetProduceListIsSubscribable];


export const GetProduceListIsSubscribable = {
  true: 'true',
  false: 'false',
} as const;